VFG9833 Master, now including the „re-function“
The „re-function“ is used to find the resonance point f.e. of a parallel LC-circuit. You connect the buffered output of the VFG9833 to your LC-circuit in series with a 100k resistor. You hook up your measure-input of the VFG to the circuit, „behind“ the resistor. Please get a rough idea of the resonance frequency. Next you define a frequency to start with your search and a frequency to stop. This is quite similar to the sweep-function.

Then you start the re-procedure with this command:     re fstart:fstop:step:mode 1 or 0

fstart is the lower frequency of the range. It`s the start frequency.
fstop is the upper frequency, where the sweep stops.
step is your frequency step. take bigger steps unless you are quite sure, that you got over the resonance point. take smaller steps with your second approach to find the point more exactly.
mode 1 means, you want a verbose output
mode 0 means, you get a comma-separated output. That can be used for further calculation or grafics.